docs, tests, lock-free counters, COW observers, rate-limiter integration
- Merge PERFORMANCE.md and JOURNAL_COMPACTION.md into ARCHITECTURE.md (22 sections);
  add threading model, synchronization design, lock-free inventory, trade-offs,
  rate-limiter design, performance engineering, benchmark results, journal compaction
- Remove docs/PERFORMANCE.md and docs/JOURNAL_COMPACTION.md (superseded)
- README: add Requirements compliance table, expand What is implemented into
  named subsections, update test counts to 125, fix test runner paths
- order_gateway: idempotent_replays_, reconciliations_, alerts_ -> atomic<uint64_t>
  with fetch_add outside operational_mutex_; order_observers_ -> atomic<shared_ptr<const
  OrderObserverList>> (COW); next_observer_token_ -> atomic<ObserverToken>
- simulated_exchange_adapter: add throw_on_place config flag and
  synchronize_rate_limiter() method
- test_rate_limiter: add 7 new gateway/observer/concurrency tests (125 total);
  fix burst sizing (burst=2 for query_balances + place)
- .gitignore: exclude evidence/, lsan.supp, include/abex/bootstrap/config_loader.hpp

+++ include/abex/domain/decimal.hpp
@@ -52,7 +52,7 @@ public:
             any_digit = true;
             if (decimal_point) {
                 if (++fraction_digits > precision) throw "decimal literal has excess precision";
-                fraction = fraction * 10 + (character - '0');
+                fraction = fraction * 10 + static_cast<std::uint64_t>(character - '0');
             } else {
                 const auto digit = static_cast<std::uint64_t>(character - '0');
                 if (whole > (raw_limit - digit) / 10U) {
